@@ -1,8 +1,28 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
+import warnings
 
 import vuelos_pb2 as vuelos__pb2
+
+GRPC_GENERATED_VERSION = '1.73.1'
+GRPC_VERSION = grpc.__version__
+_version_not_supported = False
+
+try:
+    from grpc._utilities import first_version_is_lower
+    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
+except ImportError:
+    _version_not_supported = True
+
+if _version_not_supported:
+    raise RuntimeError(
+        f'The grpc package installed is at version {GRPC_VERSION},'
+        + f' but the generated code in vuelos_pb2_grpc.py depends on'
+        + f' grpcio>={GRPC_GENERATED_VERSION}.'
+        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
+        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
+    )
 
 
 class VuelosStub(object):
@@ -22,12 +42,12 @@ class VuelosStub(object):
                 '/vuelos.Vuelos/CrearReserva',
                 request_serializer=vuelos__pb2.Reserva.SerializeToString,
                 response_deserializer=vuelos__pb2.RespuestaReserva.FromString,
-                )
+                _registered_method=True)
         self.ConsultarReserva = channel.unary_unary(
                 '/vuelos.Vuelos/ConsultarReserva',
                 request_serializer=vuelos__pb2.QueryReserva.SerializeToString,
                 response_deserializer=vuelos__pb2.RespuestaReserva.FromString,
-                )
+                _registered_method=True)
 
 
 class VuelosServicer(object):
@@ -66,6 +86,7 @@ def add_VuelosServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'vuelos.Vuelos', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
+    server.add_registered_method_handlers('vuelos.Vuelos', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -87,11 +108,21 @@ class Vuelos(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(request, target, '/vuelos.Vuelos/CrearReserva',
+        return grpc.experimental.unary_unary(
+            request,
+            target,
+            '/vuelos.Vuelos/CrearReserva',
             vuelos__pb2.Reserva.SerializeToString,
             vuelos__pb2.RespuestaReserva.FromString,
-            options, channel_credentials,
-            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
+            options,
+            channel_credentials,
+            insecure,
+            call_credentials,
+            compression,
+            wait_for_ready,
+            timeout,
+            metadata,
+            _registered_method=True)
 
     @staticmethod
     def ConsultarReserva(request,
@@ -104,8 +135,18 @@ class Vuelos(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(request, target, '/vuelos.Vuelos/ConsultarReserva',
+        return grpc.experimental.unary_unary(
+            request,
+            target,
+            '/vuelos.Vuelos/ConsultarReserva',
             vuelos__pb2.QueryReserva.SerializeToString,
             vuelos__pb2.RespuestaReserva.FromString,
-            options, channel_credentials,
-            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
+            options,
+            channel_credentials,
+            insecure,
+            call_credentials,
+            compression,
+            wait_for_ready,
+            timeout,
+            metadata,
+            _registered_method=True)
